requirements.txt
# Core dependencies for Creative Writing with KoboldCpp
pyyaml>=6.0
omegaconf>=2.3.0
rich>=13.0.0

# HTTP requests for KoboldCpp API
requests>=2.31.0

# OpenAI API for LLM-as-a-Judge evaluation
openai>=1.0.0
python-dotenv>=1.0.0

# Scientific computing (minimal)
numpy>=1.24.0

# Optional: For advanced analysis (can be removed if not needed)
# torch>=2.0.0 